+++ domain/src/main/kotlin/com/wespot/common/NotificationTypeToUpdateFeatureContent.kt
@@ -1,6 +1,6 @@
 package com.wespot.common
 
-import com.wespot.common.link.DeepLink
+import com.wespot.view.button.link.DeepLink
 import com.wespot.common.view.OnClickActionType
 import com.wespot.exception.CustomException
 import com.wespot.exception.ExceptionView
