+++ domain/src/main/kotlin/com/wespot/view/button/ButtonListComponent.kt
@@ -1,4 +1,4 @@
-package com.wespot.view
+package com.wespot.view.button
 
 data class ButtonListComponent(
     val type: String,
