+++ domain/src/main/kotlin/com/wespot/view/image/ImageComponent.kt
@@ -1,4 +1,4 @@
-package com.wespot.view
+package com.wespot.view.image
 
 data class ImageComponent(
     val type: String,
