+++ domain/src/main/kotlin/com/wespot/view/text/TextLineComponent.kt
@@ -1,4 +1,4 @@
-package com.wespot.view
+package com.wespot.view.text
 
 data class TextLineComponent(
     val icon: String,
